+++ packages/ui-connector/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @orderly.network/ui-connector
 
+## 2.8.10
+
+### Patch Changes
+
+- publish
+- 0be13b3: publish
+- Updated dependencies
+- Updated dependencies [0be13b3]
+  - @orderly.network/react-app@2.8.10
+  - @orderly.network/hooks@2.8.10
+  - @orderly.network/i18n@2.8.10
+  - @orderly.network/types@2.8.10
+  - @orderly.network/ui@2.8.10
+  - @orderly.network/ui-chain-selector@2.8.10
+
+## 2.8.10-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/react-app@2.8.10-alpha.0
+  - @orderly.network/hooks@2.8.10-alpha.0
+  - @orderly.network/i18n@2.8.10-alpha.0
+  - @orderly.network/types@2.8.10-alpha.0
+  - @orderly.network/ui@2.8.10-alpha.0
+  - @orderly.network/ui-chain-selector@2.8.10-alpha.0
+
 ## 2.8.9
 
 ### Patch Changes